Rename krg_mgp to mgp
