+++ smt/surrogate_models/__init__.py
@@ -5,7 +5,7 @@ from .krg import KRG
 from .gekpls import GEKPLS
 from .kplsk import KPLSK
 from .genn import GENN
-from .krg_mgp import MGP
+from .mgp import MGP
 
 try:
     from .idw import IDW
